mestizo-crm/backend/requirements.txt
# Django Backend Dependencies
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
drf-spectacular>=0.27,<1.0
psycopg[binary]>=3.1,<4.0
gunicorn>=21.0,<22.0
python-decouple>=3.8,<4.0

# Development & Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
